Move state.rs into crdt module
